Added rule to scale larger shadow texture to enable sharper textures with increases to resolution.
Discovered a larger shadow texture 960x3840, which turns out relates to the size of the 960x960 shadow textures. Thus texture size must scale as follows; width is always x1 and height must be x4. This enables shadows and also increases the resolution of them. I still cannot find a way to fix the shadow rendering and I currently believe this to be a issue with Cemu.